requirements.txt
pandas==1.5.3
numpy==1.23.5
requests==2.28.1
plotly==5.13.0
streamlit==1.20.0
matplotlib==3.6.3
streamlit_keplergl
numerize==0.12
pillow==9.4.0